Add a FcmManager to handle setup and message
